requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.46
asyncpg>=1.0.0
alembic>=1.18.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.2.1
pydantic>=2.12.5
bcrypt>=5.0.0
psycopg2-binary>=2.9.6